# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express API on Vercel",
    "version": "1.0.0",
    "description": "API documentation for Express.js application deployed on Vercel with OAuth 2.0 authentication",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://express-js-on-vercel-iota-seven-72.vercel.app",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token obtained from /oauth/token endpoint"
      },
      "OAuth2": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "/oauth/token",
            "scopes": {
              "read": "Read access",
              "write": "Write access"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/oauth/token": {
      "post": {
        "summary": "OAuth token endpoint",
        "description": "Generate an access token for API authentication",
        "tags": [
          "OAuth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "grant_type",
                  "client_id",
                  "client_secret"
                ],
                "properties": {
                  "grant_type": {
                    "type": "string",
                    "example": "client_credentials"
                  },
                  "client_id": {
                    "type": "string",
                    "example": "your-client-id"
                  },
                  "client_secret": {
                    "type": "string",
                    "example": "your-client-secret"
                  },
                  "scope": {
                    "type": "string",
                    "example": "read write"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "token_type": {
                      "type": "string",
                      "example": "Bearer"
                    },
                    "expires_in": {
                      "type": "number",
                      "example": 3600
                    },
                    "scope": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/protected": {
      "get": {
        "summary": "Protected API endpoint",
        "description": "Example of a protected endpoint that requires OAuth authentication",
        "tags": [
          "API"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with user data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "This is protected data"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "sub": {
                          "type": "string"
                        },
                        "scope": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Access token required"
          },
          "403": {
            "description": "Invalid or expired token"
          }
        }
      }
    }
  },
...
}